added preprocessing,scraper and summariser service

+++ backend/app/models/schemas.py
@@ -7,7 +7,7 @@ class ReviewRequest(BaseModel):
 
 class SingleReview(BaseModel):
     review_text: str = Field(...,description="Text content of the review")
-    rating: Optional[int] = Field(None,ge=1,le=5,description="Product rating from 1-5")
+    rating: Optional[float] = Field(None,ge=1,le=5,description="Product rating from 1-5")
     review_date: Optional[date] = Field(None,description="Date when the review was posted")
 
 class AnalyzeRequest(BaseModel):
